requirements.txt
pandas
psycopg
python-dotenv
instructor
sentence_transformers
torch
uvicorn
fastapi
pydantic